refactor(tetris): 重构俄罗斯方块游戏界面和交互
- 更新了 TetrisBlock 组件的样式，添加了新的 CSS 样式文件
- 重构了 TetrisGame 组件，优化了游戏控制按钮的样式和交互
- 调整了游戏页面的组件引用，使用新的 TetrisGame 组件
- 更新了首页内容组件中的 TetrisBlock 引用
