+++ src/components/solid/home-block/tetris-block/index.css
@@ -1,0 +1,50 @@
+.app-card {
+	position: relative;
+	@apply size-full;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	padding: 12px;
+	gap: 12px;
+	border-radius: 1.5rem;
+	cursor: pointer;
+	transition: all 0.6s cubic-bezier(0.175, 0.885, 0.32, 1.275);
+}
+
+.app-card::before {
+	content: "";
+	position: absolute;
+	inset: 0;
+	margin: auto;
+	border-radius: 1.5rem;
+	background: linear-gradient(-45deg, #e81cff 0%, #40c9ff 100%);
+	pointer-events: none;
+	transition: all 0.6s cubic-bezier(0.175, 0.885, 0.32, 1.275);
+}
+
+.app-card::after {
+	content: "";
+	z-index: -1;
+	position: absolute;
+	inset: 0;
+	background: linear-gradient(-45deg, #fc00ff 0%, #00dbde 100%);
+	transform: translate3d(0, 0, 0) scale(0.9);
+	filter: blur(20px);
+}
+
+.app-card p:last-child {
+	color: #e81cff;
+	font-weight: 600;
+}
+
+.app-card:hover::after {
+	filter: blur(30px);
+}
+
+.app-card:hover::before {
+	/* transform: rotate(-90deg) scale(1.2); */
+}
+
+.app-card:hover {
+	transform: rotate(360deg) scale(0.8);
+}
