+++ src/pages/games/tetris/index.astro
@@ -1,6 +1,6 @@
 ---
 import Layout from "../../../layouts/Layout.astro";
-import TetrisGame from "../../../components/solid/tetris";
+import TetrisGame from "../../../components/solid/tetris/tetris";
 ---
 <Layout pageTitle="俄罗斯方块">
  <TetrisGame client:load  />
