+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
 import { cn } from "../utils";
 import { createSignal, For } from "solid-js";
 import { BlockList, TabsObj, type TabsKey } from "../common/index";
-import TetrisGame from "../components/solid/tetris";
+import TetrisGame from "../components/solid/tetris/tetris";
 
 const Content = () => {
   const [active, setActive] = createSignal<TabsKey>("all");
